Move reqwest client to subdirectory
This change moves the ScabbardClient implementation (and builder) to a
reqwest submodule.  As of this commit, it is re-exported from the base
client module.

The return values remain in the root client module, as these will be
used in a trait to be extracted in a future commit.
